+++ yaml-config-updater/src/test/groovy/ru/vyarus/yaml/updater/UpdateFlowPropsTest.groovy
@@ -50,7 +50,7 @@ addition2: {t: t}
 """
         and: "report correct"
         print(report, curSize, current.size()) == """Configuration: /tmp/CONFIG.yml (300 bytes, 21 lines)
-Updated from source of 255 bytes, 27 lines
+Updated from source of 256 bytes, 27 lines
 Resulted in 301 bytes, 25 lines
 
 \tAdded from new file: